remove unused files, rename AssetList and SketchList to AssetListRow, SketchListRow, move CollectionListRow.jsx to ../
