+++ client/modules/IDE/components/AssetListRow.jsx
