--- client/modules/IDE/components/CollectionList/index.js
@@ -1,1 +1,0 @@
-export { default } from './CollectionList';
