+++ client/modules/IDE/components/CollectionListRow.jsx
@@ -4,13 +4,13 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router';
 import { bindActionCreators } from 'redux';
 import { withTranslation } from 'react-i18next';
-import * as ProjectActions from '../../actions/project';
-import * as CollectionsActions from '../../actions/collections';
-import * as IdeActions from '../../actions/ide';
-import * as ToastActions from '../../actions/toast';
-import dates from '../../../../utils/formatDate';
+import * as ProjectActions from '../actions/project';
+import * as CollectionsActions from '../actions/collections';
+import * as IdeActions from '../actions/ide';
+import * as ToastActions from '../actions/toast';
+import dates from '../../../utils/formatDate';
 
-import DownFilledTriangleIcon from '../../../../images/down-filled-triangle.svg';
+import DownFilledTriangleIcon from '../../../images/down-filled-triangle.svg';
 
 class CollectionListRowBase extends React.Component {
   static projectInCollection(project, collection) {
